Restructured the unit test files
