+++ app/api/v1/incidents/incident_models.py
@@ -1,9 +1,5 @@
 import psycopg2
 import os
-"""
-url = "dbname='ireporter' host='localhost'\
-            port='5432' user='postgres' password='Nanbada13'"
-"""
 DATABASE_URL = os.getenv('DATABASE_URL')
